requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
aiosqlite==0.20.0
alembic==1.14.1
amqp==5.3.1
anyio==4.8.0
apache-airflow==2.10.4
apache-airflow-providers-common-compat==1.3.0
apache-airflow-providers-common-io==1.5.0
apache-airflow-providers-common-sql==1.21.0
apache-airflow-providers-fab==1.5.2
apache-airflow-providers-ftp==3.12.0
apache-airflow-providers-http==5.0.0
apache-airflow-providers-imap==3.8.0
apache-airflow-providers-smtp==1.9.0
apache-airflow-providers-sqlite==4.0.0
apache-superset==4.1.1
apispec==6.8.1
apsw==3.48.0.0
argcomplete==3.5.3
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.3.0
babel==2.16.0
backoff==2.2.1
bcrypt==4.2.1
billiard==4.2.1
blinker==1.9.0
Bottleneck==1.4.2
Brotli==1.1.0
cachelib==0.9.0
cachetools==5.5.1
cattrs==24.1.2
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-option-group==0.5.6
click-plugins==1.1.1
click-repl==0.3.0
clickclick==20.10.2
colorama==0.4.6
colorlog==6.9.0
ConfigUpdater==3.2
connexion==2.14.2
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==42.0.8
Deprecated==1.2.15
deprecation==2.1.0
dill==0.3.9
dnspython==2.7.0
email_validator==2.2.0
exceptiongroup==1.2.2
Flask==2.2.5
Flask-AppBuilder==4.5.2
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-Compress==1.17
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Migrate==3.1.0
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
flask-talisman==1.1.0
Flask-WTF==1.2.2
frozenlist==1.5.0
fsspec==2024.12.0
func_timeout==4.3.5
geographiclib==2.0
geopy==2.4.1
gevent==24.11.1
google-auth==2.37.0
google-re2==1.1.20240702
googleapis-common-protos==1.66.0
greenlet==3.1.1
grpcio==1.70.0
gunicorn==23.0.0
h11==0.14.0
hashids==1.3.1
holidays==0.25
httpcore==1.0.7
httpx==0.28.1
humanize==4.11.0
idna==3.10
ijson==3.3.0
importlib_metadata==8.5.0
inflection==0.5.1
isodate==0.7.2
itsdangerous==2.2.0
Jinja2==3.1.5
jmespath==1.0.1
jpype1==1.5.1
jsonpath-ng==1.7.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.4.2
korean-lunar-calendar==0.3.1
lazy-object-proxy==1.10.0
limits==4.0.1
linkify-it-py==2.0.3
llvmlite==0.43.0
lockfile==0.12.2
Mako==1.3.8
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
more-itertools==10.6.0
msgpack==1.0.8
msgspec==0.19.0
multidict==6.1.0
nh3==0.2.20
numba==0.60.0
numexpr==2.10.2
numpy==1.23.5
opentelemetry-api==1.29.0
opentelemetry-exporter-otlp==1.29.0
opentelemetry-exporter-otlp-proto-common==1.29.0
opentelemetry-exporter-otlp-proto-grpc==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
opentelemetry-proto==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-semantic-conventions==0.50b0
ordered-set==4.1.0
outcome==1.3.0.post0
packaging==24.2
pandas==2.0.3
paramiko==3.5.0
parsedatetime==2.6
pathspec==0.12.1
pendulum==3.0.0
pgsanity==0.2.9
pillow==11.1.0
platformdirs==4.3.6
pluggy==1.5.0
ply==3.11
polyline==2.0.2
prison==0.2.1
prompt_toolkit==3.0.50
propcache==0.2.1
protobuf==5.29.3
psutil==6.1.1
psycopg2-binary==2.9.10
pyarrow==14.0.2
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
Pygments==2.19.1
PyJWT==2.10.1
pymongo==4.10.1
PyNaCl==1.5.0
pyodbc==5.2.0
pyOpenSSL==25.0.0
pyparsing==3.2.1
PySocks==1.7.1
python-daemon==3.1.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-geohash==0.8.5
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.2
PyYAML==6.0.2
redis==4.6.0
referencing==0.36.1
requests==2.32.3
requests-cache==1.2.1
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.6.0
rpds-py==0.22.3
rsa==4.9
selenium==4.9.1
setproctitle==1.3.4
shillelagh==1.3.4
shortid==0.1.2
simplejson==3.19.3
six==1.17.0
slack_sdk==3.34.0
sniffio==1.3.1
sortedcontainers==2.4.0
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.38.3
sqlalchemy_drill==1.1.5
sqlglot==25.34.1
sqlparse==0.5.3
sshtunnel==0.4.0
tabulate==0.8.10
tenacity==9.0.0
termcolor==2.5.0
text-unidecode==1.3
time-machine==2.16.0
trio==0.28.0
trio-websocket==0.11.1
typing_extensions==4.12.2
tzdata==2025.1
uc-micro-py==1.0.3
universal_pathlib==0.2.6
url-normalize==1.4.3
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.2
wsproto==1.2.0
WTForms==3.2.1
WTForms-JSON==0.3.5
XlsxWriter==3.0.9
yarl==1.18.3
zipp==3.21.0
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
